Add ref check in organization chart upload.

+++ controls/organization/upload.php
@@ -3,20 +3,22 @@
 
 use Intra\Model\LightFileModel;
 
-$infile = $_FILES["fileToUpload"]["tmp_name"];
+if ($_FILES && $_FILES["fileToUpload"]) {
+	$infile = $_FILES["fileToUpload"]["tmp_name"];
 
-if ($infile) {
-	$filebag = new LightFileModel('organization');
-	$outfile = $filebag->getLocation(date("Y-m-d") . ".pdf");
+	if ($infile) {
+		$filebag = new LightFileModel('organization');
+		$outfile = $filebag->getLocation(date("Y-m-d") . ".pdf");
 
-	if (!move_uploaded_file($infile, $outfile)) {
-		die('파일을 업로드하지 못했습니다.');
+		if (!move_uploaded_file($infile, $outfile)) {
+			die('파일을 업로드하지 못했습니다.');
+		}
+
+		$recent = $filebag->getLocation('recent');
+
+		unlink($recent);
+		symlink($outfile, $recent);
 	}
-
-	$recent = $filebag->getLocation('recent');
-
-	unlink($recent);
-	symlink($outfile, $recent);
 }
 
 return [];
